requirements.txt
flet==0.24.1
sqlalchemy==2.0.23
bcrypt==4.1.1
python-dotenv==1.0.0
requests==2.31.0
reportlab==4.0.7
psycopg2-binary==2.9.9
